review: dedupe is_no_peers_subscribed_error, clarify governance asymmetry
Addresses bot review feedback on #2497:

* **Bugbot — duplicated NoPeersSubscribed detection.** Move
  `is_no_peers_subscribed_error` from `calimero-node-primitives` into
  `calimero-network-primitives::client` so `calimero-context`'s
  `classify_network_publish_error` can delegate to the same helper.
  Both downstream crates already depend on `calimero-network-primitives`
  and gossipsub is already a feature there, so no new deps. Helper
  tests move with it.
* **MeroReviewer + AI Code Reviewer — publish_on_namespace asymmetry.**
  Add doc comments to `publish_on_namespace` and
  `publish_on_namespace_now` explaining why they keep the mesh-peer
  wait and propagate `NoPeersSubscribedToTopic` while
  `broadcast` / `broadcast_heartbeat` silence it: governance ops have
  no sync-pull recovery path (per #2369 post-mortem), so the
  publisher-side wait + hard-fail-on-no-peers is the correct contract.
  Behavior unchanged; future reviewers won't flag this as a missed fix.
* **MeroReviewer nitpick — heartbeat log wording.** "Next tick will
  retry" implies re-sending the same message; replace with "next tick
  will carry fresh state" to match the actual semantics.
* **MeroReviewer suggestion — eyre-chain doc note.** Add a paragraph to
  `is_no_peers_subscribed_error`'s doc comment clarifying that it only
  walks `eyre::Report::chain()`. Documents the scope; no behavior change.

All callers compile against the moved helper. Suites green:
calimero-network-primitives 14/14, calimero-node-primitives 323/323,
calimero-context 373/373.
